+++ src/components/common/Uploader/UploaderContainer.js
@@ -1,9 +1,8 @@
 // Uploader logic
-
 import React, { useState } from 'react';
 import RenderUploader from './RenderUploader';
 import { useLocalStorage } from '../../../utils/hooks';
-import { LoadingComponent } from '../../common';
+import { LoadingComponent } from '..';
 import { useHistory } from 'react-router-dom';
 
 import { message } from 'antd';
@@ -11,14 +10,13 @@ import { message } from 'antd';
 // api
 import { uploadSubmissionData, getData } from '../../../api';
 
-const Uploader = () => {
+const Uploader = ({ fileLimit, uploadURL }) => {
   const [fileList, setFileList] = useState([]);
   const [isLoading, setIsLoading] = useState(false);
   const [errorState, setErrorState] = useState(false);
 
   const [userId] = useLocalStorage('curUserId');
   const [curUserToken] = useLocalStorage('curUserToken');
-  let missionId = '';
 
   const { push } = useHistory();
 
@@ -63,7 +61,7 @@ const Uploader = () => {
     const formData = new FormData();
 
     // 'child/userId/mission'
-    const endpoint = `child/${userId}/mission/write`;
+    const endpoint = uploadURL;
     fileList.forEach(file => {
       formData.append('images', file.originFileObj);
     });
@@ -90,6 +88,7 @@ const Uploader = () => {
       <RenderUploader
         errorState={errorState}
         fileList={fileList}
+        fileLimit={fileLimit}
         onChange={onChange}
         onPreview={onPreview}
         onSubmit={onSubmit}
